requirements.txt
Cython==0.27.3
llvmlite==0.20.0
lxml==3.5.0
matplotlib==2.1.0
numba==0.35.0
numpy==1.13.3
pathlib2==2.3.0
Pillow==4.3.0
scipy==1.0.0
stop-words==2015.2.23.1
tensorflow-gpu==1.3.0
tensorflow-tensorboard==0.1.8
xmltodict==0.10.2
opencv-python==3.3.0.10